requirements.txt
PyYAML==5.1.2
tabulate==0.8.3
youtube-dl==2019.8.13
